make IO.writev safely interruptible and 1.9-thread-safe
Since writev() may be used for "slow" devices where writes may
be interrupted from signals, we need to ensure the interpreter
is in a sane state under 1.8 using TRAP_BEG/TRAP_END.  Under
1.9, we can release the GVL and just let the thread block as
long as it needs to for the writev to complete.

Keep in mind that our thread will still block all threads under
MRI 1.8, but 1.8 is becoming less and less important.