writev: retry on partial writes
Since writev(2) may be (and should be often) used on "slow" file
descriptors which may be interrupted, make this method easier to use
and retry for the user.  This also has the pleasant side effect
of making writev compatible with non-blocking green threads
under MRI 1.8 (assuming the descriptor already has the
O_NONBLOCK flag set).